+++ client/src/assets/home_carousel/carousel1.png
