+++ client/src/assets/home_carousel/carousel2.png
