+++ client/src/assets/home_carousel/carousel3.png
