+++ client/src/assets/route_cards/kpop-items.png
