+++ client/src/pages/Homepage.jsx
@@ -2,9 +2,9 @@ import Carousel from '../components/Carousel';
 import RouteCard from '../components/RouteCard';
 import InstaPlugin from '../components/InstaPlugIn';
 import front from '../assets/kingdom_front.png'
-import bake from '../assets/baked-item.jpg';
-import kpop from '../assets/kpop-items.png';
-import books from '../assets/books.png';
+import bake from '../assets/route_cards/baked-item.jpg';
+import kpop from '../assets/route_cards/kpop-items.png';
+import books from '../assets/route_cards/books.png';
 import { IoArrowForwardCircleOutline } from "react-icons/io5";
